feat(users): add suspend users

+++ app/(app)/users/_components/UsersDataTable/Menu/index.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { ArrowRightLeft, BadgeCheck, CircleOff, Edit, KeyRound, MoreHorizontal, Trash } from 'lucide-react';
+import { ArrowRightLeft, BadgeCheck, Edit, KeyRound, Lock, MoreHorizontal, Trash, Unlock } from 'lucide-react';
 import Link from 'next/link';
 import { useRouter } from 'next/navigation';
 import { signOut, useSession } from 'next-auth/react';
@@ -9,9 +9,11 @@ import { useState } from 'react';
 import { Button } from '@/components/ui/button';
 import { DropdownMenu, DropdownMenuContent, DropdownMenuItem, DropdownMenuLabel, DropdownMenuSeparator, DropdownMenuTrigger } from '@/components/ui/dropdown-menu';
 import ConfirmationFormModal, { ConfirmationModalOpenChangeEvent } from '@/components/ui/Modal/ConfirmationFormModal';
+import ConfirmationModal from '@/components/ui/Modal/ConfirmationModal';
 import { useToast } from '@/components/ui/use-toast';
 import useCsrf from '@/context/csrf/useCsrf';
-import { deleteUser } from '@/services/users.service';
+import useUsers from '@/context/users/useUsers';
+import { deleteUser, updateUser as updateUserQuery } from '@/services/users.service';
 import { ApiError, getErrorMessage } from '@/utils/error';
 
 import { UserColumn } from '../columns';
@@ -20,11 +22,35 @@ type MenuProps = {
 	rowData: UserColumn;
 }
 
+type ModalState = {
+	isOpen: boolean;
+	action: 'delete' | 'suspend' | 'activate';
+}
+
+const getModalContent = (rowData: UserColumn) => ({
+	delete: {
+		title: 'Delete user',
+		description: <span>Please enter the email of the user <span className="font-bold text-slate-700 select-none">{ rowData.email }</span> to confirm the deletion. This action is irreversible.</span>,
+	},
+	suspend: {
+		title: 'Suspend user',
+		description: <span>Please enter the email of the user <span className="font-bold text-slate-700 select-none">{ rowData.email }</span> to confirm the account suspension. The user will no longer be able to log in.</span>,
+	},
+	activate: {
+		title: 'Activate user',
+		description: <span>Please confirm the account activation. The user will be able to log in again.</span>,
+	},
+});
+
 const Menu = ({ rowData }: MenuProps) => {
 
 	const { csrfToken } = useCsrf();
+	const { updateUser } = useUsers();
 
-	const [ isConfirmationModalOpen, setIsConfirmationModalOpen ] = useState<boolean>(false);
+	const [ confirmationModalState, setConfirmationModalState ] = useState<ModalState>({
+		isOpen: false,
+		action: 'delete', 
+	});
 	const [ isLoading, setIsLoading ] = useState<boolean>(false);
 
 	const router = useRouter();
@@ -32,12 +58,32 @@ const Menu = ({ rowData }: MenuProps) => {
 	const { toast } = useToast();
 
 	const handleDeleteUser = () => {
-		setIsConfirmationModalOpen(true);
+		setConfirmationModalState({
+			isOpen: true,
+			action: 'delete', 
+		});
+	};
+
+	const handleSuspendUser = () => {
+		setConfirmationModalState({
+			isOpen: true,
+			action: 'suspend', 
+		});
+	};
+
+	const handleActivateUser = () => {
+		setConfirmationModalState({
+			isOpen: true,
+			action: 'activate', 
+		});
 	};
 
 	const handleConfirmationModalOpenChange: ConfirmationModalOpenChangeEvent = async ({ openState, isConfirmed }) => {
 		if (!isConfirmed) {
-			return setIsConfirmationModalOpen(openState);
+			return setConfirmationModalState({
+				...confirmationModalState,
+				isOpen: openState, 
+			});
 		}
 
 		if (!csrfToken) {
@@ -46,14 +92,40 @@ const Menu = ({ rowData }: MenuProps) => {
 
 		try {
 			setIsLoading(true);
-			await deleteUser(rowData.id, csrfToken);
-			setIsConfirmationModalOpen(openState);
-			if (session && rowData.id.toString() === session.user.id.toString()) {
+			if (confirmationModalState.action === 'delete') {
+				await deleteUser(rowData.id, csrfToken);
+				// TODO => Refetch users on delete
+			}
+			if (confirmationModalState.action === 'suspend') {
+				await updateUserQuery({
+					id: rowData.id,
+					is_disabled: true,
+				}, csrfToken);
+				updateUser({
+					...rowData,
+					is_disabled: true, 
+				});
+			}
+			if (confirmationModalState.action === 'activate') {
+				await updateUserQuery({
+					id: rowData.id,
+					is_disabled: false,
+				}, csrfToken);
+				updateUser({
+					...rowData,
+					is_disabled: false, 
+				});
+			}
+			setConfirmationModalState({
+				...confirmationModalState,
+				isOpen: openState,
+			});
+			if (session && rowData.id.toString() === session.user.id.toString() && [ 'delete', 'suspend' ].includes(confirmationModalState.action)) {
 				await signOut({ redirect: false });
 				return router.replace('/signin');
 			}
-			router.refresh();
 		} catch (error) {
+			console.error(error);
 			const apiError = error as ApiError<unknown>;
 			toast({
 				title: 'Error',
@@ -104,11 +176,20 @@ const Menu = ({ rowData }: MenuProps) => {
 					><ArrowRightLeft size="16" /> Impersonate
 					</DropdownMenuItem>
 					<DropdownMenuSeparator />
-					<DropdownMenuItem
-						className="gap-2 text-red-500 hover:cursor-pointer"
-						disabled
-					><CircleOff size="16" /> Suspend
-					</DropdownMenuItem>
+					{
+						rowData.is_disabled ?
+							<DropdownMenuItem
+								className="gap-2 hover:cursor-pointer"
+								onClick={ handleActivateUser }
+							><Unlock size="16" /> Activate
+							</DropdownMenuItem>
+							:
+							<DropdownMenuItem
+								className="gap-2 text-red-500 hover:cursor-pointer"
+								onClick={ handleSuspendUser }
+							><Lock size="16" /> Suspend
+							</DropdownMenuItem>
+					}
 					<DropdownMenuItem
 						className="gap-2 text-red-500 hover:cursor-pointer"
 						onClick={ handleDeleteUser }
@@ -119,12 +200,19 @@ const Menu = ({ rowData }: MenuProps) => {
 			</DropdownMenu>
 			<ConfirmationFormModal
 				data={ rowData }
-				description={ <span>Please enter the email of the user <span className="font-bold text-slate-700 select-none">{ rowData.email }</span> to confirm the deletion. This action is irreversible.</span> }
+				description={ getModalContent(rowData)[ confirmationModalState.action ].description }
 				isLoading={ isLoading }
-				isOpen={ isConfirmationModalOpen }
+				isOpen={ confirmationModalState.action !== 'activate' && confirmationModalState.isOpen }
 				keyToValidate="email"
-				title="Delete user"
+				title={ getModalContent(rowData)[ confirmationModalState.action ].title }
 				variant="destructive"
+				onOpenChange={ handleConfirmationModalOpenChange }
+			/>
+			<ConfirmationModal
+				description={ getModalContent(rowData)[ confirmationModalState.action ].description }
+				isLoading={ isLoading }
+				isOpen={ confirmationModalState.action === 'activate' && confirmationModalState.isOpen }
+				title={ getModalContent(rowData)[ confirmationModalState.action ].title }
 				onOpenChange={ handleConfirmationModalOpenChange }
 			/>
 		</>
